+++ lib/Resque/Job.php
@@ -129,8 +129,16 @@ class Resque_Job
 				'Job class ' . $this->payload['class'] . ' does not contain a perform method.'
 			);
 		}
+		
+		if(method_exists($this->payload['class'], 'setUp')) {
+			call_user_func(array($this->payload['class'], 'setUp'), $this->payload['args']);
+		}
 
 		call_user_func(array($this->payload['class'], 'perform'), $this->payload['args']);
+		
+		if(method_exists($this->payload['class'], 'tearDown')) {
+			call_user_func(array($this->payload['class'], 'tearDown'), $this->payload['args']);
+		}
 	}
 
 	/**