+++ CHANGELOG.markdown
@@ -1,6 +1,8 @@
 ## 1.1 (????-??-??) ##
 * Change arguments for jobs to be an array as they're easier to work with in
-PHP
+PHP.
+* Implement ability to have setUp and tearDown methods for jobs, called before
+and after every single run.
 
 ## 1.0 (2010-04-18) ##